LayoutTests/platform/ios-sim-deprecated/svg/animations/additive-values-width-animation-expected.txt

SVG 1.1 dynamic animation tests

This tests values animation and additive='sum'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 10
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 60
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 93.3
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 143.33
PASS rect.width.baseVal.value is 60
PASS rect.width.animVal.value is 160
PASS rect.width.baseVal.value is 60
PASS rect.width.animVal.value is 160
PASS rect.width.baseVal.value is 60
PASS successfullyParsed is true

TEST COMPLETE
